Created atlassian-webdriver module.
Seperated out the atlassian-selenium-browsers into another module atlassian-browsers.
Now the atlassian-selenium-browsers-auto uses the atlassian-browsers-auto as a dependency.

Created atlassian-webdriver-browsers-auto which also uses atlassian-browsers-auto.

Reworked the auto install logic.

Added in firefox 3.6 and chrome 5 for mac. So far have only tested with Webdriver.

Added in the atlassian-webdriver module which has webdriver-core and webdriver-jira (more workj to be done here).



Commit rewritten to remove blob.