SELENIUM-208: improved use of generics in Poller and Conditions
using Matcher<T> doesn't allow callers to pass in Matcher<U> where U is
a superclass of T, and doesn't play well with Hamcrest's core matchers.

it's better to use Matcher<? super T> in this case so that callers don't
have to jump through hoops to get the exact same generic type.